Python Scripts/Project 4 - Automating Google Sheets 2/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
gspread
oauth2client
azure-identity
azure-keyvault-secrets
azure-storage-blob